BigDecimal migration

+++ CAInvoice/src/main/java/ua/its/slot7/caccounting/service/BLService.java
@@ -10,6 +10,7 @@ import ua.its.slot7.caccounting.model.invoiceline.InvoiceLine;
 import ua.its.slot7.caccounting.model.person.Person;
 import ua.its.slot7.caccounting.model.user.User;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -118,7 +119,7 @@ public class BLService implements BLServiceAvatar {
 	}
 
 	@Override
-	public float calcInvoiceSum(Invoice invoice) {
+	public BigDecimal calcInvoiceSum(Invoice invoice) {
 		return invoice.calcInvoiceSum();
 	}
